7. Разделяй и властвуй: Добавил абстрактный класс для отрисовки компонентов, сделал наследование от этого класса для отображений компонентов
