+++ src/view/button.js
@@ -1,25 +1,9 @@
-import {createElement} from "../utils";
+import Abstract from "./abstract";
 
 const createButtonTemplate = () => `<button class="films-list__show-more">Show more</button>`;
 
-export default class LoadMoreButton {
-  constructor() {
-    this._element = null;
-  }
-
+export default class LoadMoreButton extends Abstract {
   _getTemplate() {
     return createButtonTemplate();
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
