+++ src/view/films-container.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import Abstract from "./abstract";
 
 const createFilmsContainerTemplate = () =>
   `<section class="films">
@@ -16,24 +16,8 @@ const createFilmsContainerTemplate = () =>
     </section>
   </section>`;
 
-export default class FilmsContainer {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FilmsContainer extends Abstract{
   _getTemplate() {
     return createFilmsContainerTemplate();
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
