+++ src/view/footer-statistics.js
@@ -1,22 +1,14 @@
-import {createElement} from "../utils";
+import Abstract from "./abstract";
 
 const createFooterStatisticsTemplate = (totalFilms) => `<p>${totalFilms} movies inside</p>`;
 
-export default class FooterStatistics {
+export default class FooterStatistics extends Abstract {
   constructor(totalFilms) {
-    this._element = null;
+    super();
     this._totalFilms = totalFilms;
   }
 
   _getTemplate() {
     return createFooterStatisticsTemplate(this._totalFilms);
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
   }
 }
