+++ src/view/no-film.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import Abstract from "./abstract";
 
 const createNoFilmsTemplate = () =>
   `<section class="films">
@@ -7,24 +7,8 @@ const createNoFilmsTemplate = () =>
     </section>
   </section>`;
 
-export default class NoFilm {
-  constructor() {
-    this._element = null;
-  }
-
+export default class NoFilm extends Abstract {
   _getTemplate() {
     return createNoFilmsTemplate();
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
